+++ swift_browser_ui_frontend/currents.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  projectId: "sd-connect",
+  recordKey: process.env.CYPRESS_API_KEY,
+  cloudServiceUrl: process.env.CYPRESS_CLOUD_URL,
+};